Initial 0.6 Update
This version of the code is focused on making the getting the json from an api work in json. Then this json is stringified into an actual JSON file (Creates one if none there). Afterwards we need to make sure that the custom wordpress table block is working and parses through the JSON file!